creating TeamsTable
