+++ src/pages/Teams/components/TeamsTable/style.js
@@ -1,0 +1,9 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStylesLocal = makeStyles({
+  table: {
+    minWidth: 250,
+  },
+});
+
+export default useStylesLocal;
