--- src/pages/Teams/style.js
@@ -1,9 +1,0 @@
-import { makeStyles } from '@material-ui/core/styles';
-
-const useStylesLocal = makeStyles({
-  table: {
-    minWidth: 250,
-  },
-});
-
-export default useStylesLocal;
